chore: move internal stuff to the same file as public stuff
